+++ ApiClient.js
@@ -9,6 +9,7 @@ export default class ApiClient {
 				return responseJSON;
 			})
 			.catch(err => {
+				err = err.replace('<', '').replace('>', '');
 				console.error(err);
 			});
 	}
